Cart Tests Done

+++ src/main/java/com/glc/alisnobbacartservice/Controller/CartController.java
@@ -1,8 +1,10 @@
 package com.glc.alisnobbacartservice.Controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -36,8 +38,16 @@ public class CartController {
         return cartRepository.findById(id).orElse(null);
     } 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id){
-        cartRepository.deleteById(id);
+    public ResponseEntity<Void> delete(@PathVariable Long id){
+        Optional<CartModel> c = cartRepository.findById(id);
+
+        if(c.isPresent()){
+            cartRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        else{
+            return ResponseEntity.notFound().build();
+        }
     }
     @DeleteMapping("/deleteall")
     public void deleteAll(){
